+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
--
+- Graceful admin notice when `vendor/autoload.php` is missing instead of fatal error
 
 ## [0.2.0] - 2026-03-15